requirements.txt
mss>=9.0.1
Pillow>=10.0.0
numpy>=1.26.0
psutil>=5.9.0

pygetwindow>=0.0.9 ; sys_platform == "win32"
pywin32>=306 ; sys_platform == "win32"

pyobjc-framework-Quartz>=10.0 ; sys_platform == "darwin"
